REUSABLE: -Added disabled style to the button.
TREEVIEW: -Implemented sending one event by triggering a 'SEND_CHANGES' action inside mouseEvent before resetting the bubble key when no mouse buttons are being held down. The 'SEND_CHANGES' dispatch is being sent as an argument inside OnSaveScheduler, so that the table can also trigger a 'SEND_CHANGES' action. Inside the epic.js, it gets the changes after validating them, and if it contains a sendChanges sent inside the action, it meants the 'TRY_BUBBLE_TRANFORM' came from the table so it dispatches the sendChanges with the changes.